fix: validator

+++ Isu/Entities/Group.cs
@@ -26,7 +26,8 @@ namespace Isu.Entities
         private bool IsGroupNameCorrect(string name)
         {
             if (name.Length != 5) throw new IsuException("Name of the group does not follow certain rules. TIP: M3XYY");
-            return name[0] == 'M' && name[1] == '3' && name[2] > 0 && name[2] < 5 && name[3] >= 0 && name[3] <= 9 && name[4] >= 0 && name[4] <= 9;
+            return name[0] == 'M' && name[1] == '3' && (int)name[2] > 0 && (int)name[2] < 5 && (int)name[3] >= 0 && (int)name[3] <= 9 &&
+                   (int)name[4] >= 0 && (int)name[4] <= 9;
         }
     }
 }
